requirements.txt
Django>=2.2
psycopg2-binary==2.8.4
python-dateutil==2.6.1
coverage==4.5
tblib==1.3.2
raven==6.10.0
django-cors-headers
djangorestframework-jwt
dj-rest-auth
django-referrer-policy